An experiment to try to speed up the IMDCT step.
The strategy was:
1. Use a method from Kidambi 1998 to calculate DCT-IV.
2. Transform the result of DCT-IV to an IMDCT.

The Kidambi method still uses many inner-loop multiplications and isn't a
speed win. I think it is better-suited for hardware implementation than
software.

It is still an improvement in that it doesn't require a giant 8.5 megabyte
lookup table.

It's possible that I could optimize this further, but I should look elsewhere
for fast DCT-IV or IMDCT algorithms.